+++ tests_integration/spcs/test_cp.py
@@ -39,7 +39,7 @@ def test_cp(runner, snowflake_session):
     assert contains_row_with(result.json, row_from_snowflake_session(expect)[0])
 
     result = runner.invoke_with_connection_json(
-        ["spcs", "compute-pool", "stop", cp_name]
+        ["spcs", "compute-pool", "stop-all", cp_name]
     )
     assert contains_row_with(
         result.json,